:hammer: Merge the MobSF scanner
